requirements.txt
Django>=4.2,<5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
django-cors-headers==4.3.1
whitenoise==6.7.0

# Producción Linux (Railway/Render) — descomentar en el servidor:
# gunicorn==22.0.0
# psycopg2-binary==2.9.9
# dj-database-url==2.2.0

